requirements/base.txt
Django~=3.2.5
django-appconf~=1.0.5
django-ckeditor~=6.7.1
django-cleanup~=6.0.0
django-compressor~=4.3.1
django-crispy-forms~=1.14.0
django-crum~=0.7.9
django-debug-toolbar~=4.2.0
django-extensions~=3.2.3
django-filter~=21.1
django-mathfilters~=1.0.0
django-multiselectfield~=0.1.12
django-multiupload~=0.6.1
django-phonenumber-field~=7.0.0
django-select2~=8.1.0
# django-silk~=5.0.3
django-storages~=1.13.2
environs~=9.5.0
inflect~=6.0.2
matplotlib~=3.7.1
numpy~=1.24.2
openpyxl~=3.1.0
pandas~=1.5.3
pdfplumber~=0.10.3
phonenumbers~=8.13.0
Pillow~=9.3.0
plotly~=5.7.0
psycopg2-binary~=2.9.5
pyparsing~=3.0.8
PyPDF2~=3.0.1
python-dateutil~=2.8.2
python-dotenv~=0.21.1
pytz~=2022.7.1
reportlab~=3.6.12
sqlparse~=0.4.3
seaborn~=0.12.2
tzlocal~=5.0.1
whitenoise~=6.1.0
